pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omnilint"
version = "0.1.0"
description = "AI Dataset Quality Auditor - Tabular & Image"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Audtic Nagata"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas",
    "numpy",
    "scipy",
    "scikit-learn",
    "faiss-cpu",
    "sentence-transformers",
    "typer",
    "streamlit",
    "rich",
    "jinja2",
    "plotly",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "mypy",
    "build",
    "twine",
]
image = [
    "Pillow",
    "opencv-python",
    "imagehash",
    "torch",
    "clip",
]

[project.scripts]
omnilint = "cli.main:app"
omnilint-ui = "cli.ui:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["auditor*", "cli*", "app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88